requirements.txt
openai
pydantic
python-dotenv
requests
uvicorn
fastapi
kali_driver
kali-driver
docker
paramiko